Added check for handling retina display scaling and check for ipad/iphone resolution
Previously bar code recognition did not work on iPhone 4s as code
checked only for iPhone 4. Reimplemented this code to be iPhone
hardware agnostic, only checks if device is iPhone or iPad and if has
retina display or not. iPad constants probably need to be modified as I
don't have an iPad to test with.